+++ Genius/Views/Settings/UpdatesSection.swift
@@ -33,11 +33,7 @@ struct UpdatesSection: View {
 						updater.automaticallyChecksForUpdates = newValue != .disabled
 						updater.automaticallyDownloadsUpdates = newValue == .enabled
 					}
-				if
-					!Defaults.Keys.betaUpdates.isDefaultValue ||
-					Defaults[.developmentMode] ||
-					Defaults[.interfaceMode] >= .advanced
-				{
+				if !Defaults.Keys.betaUpdates.isDefaultValue || Defaults[.interfaceMode] >= .advanced {
 					SettingToggle(.betaUpdates, value: $betaUpdates, key: .betaUpdates)
 				}
 			}